requirements.txt
django==3.2.12
djangorestframework==3.13.1
django-filter==21.1
drf-spectacular==0.22.0